+++ mindsdb/integrations/handlers/file_handler/requirements.txt
@@ -2,3 +2,4 @@ charset-normalizer
 python-magic >= 0.4.27
 openpyxl  # Optional dep of langchain for txt files
 pymupdf   # Optional dep for PDF files
+langchain==0.0.347
