+++ mindsdb/integrations/handlers/langchain_embedding_handler/requirements.txt
@@ -1,2 +1,3 @@
 openai == 1.6.1
 tiktoken~=0.4.0
+langchain==0.0.347
