+++ mindsdb/integrations/handlers/rag_handler/requirements.txt
@@ -7,3 +7,4 @@ writerai~=1.1.0
 pydantic
 sentence-transformers
 faiss-cpu
+langchain==0.0.347
