+++ requirements/requirements.txt
@@ -34,6 +34,5 @@ gunicorn
 scikit-learn==1.3.2
 protobuf==3.20.3
 hierarchicalforecast~=0.4.0
-langchain==0.0.347
 google-auth-oauthlib
 msal
